Removes threading from the binance client.
Uses consistent async calls from within the library. This will
allow the ctrl+c control flow of the library to work again as
well as be a cleaner implementation for requests.

Switched out usage of requests with treq, which is the twisted
port of requests.

Uses twisted.internet.defer.inlineCallbacks to facilitate pseudo
blocking IO with yield statements. This should be much better
performance-wise and also be easier to debug down the road.

@inlineCallbacks decorators are fucking awesome!