Updated the function description and changed the for loop to single statement
Changed the function description to state that the return type is float. The function raises the same exception as other functions and hence the element not found is handled through that.

Tha for loop is changed to a single statement which makes more sense and the price is extracted from that.